restructure of the project library
