+++ src/models/Thought.js
