+++ src/routes/thoughts.js
